refactor : Ladder가 아닌 Line에서 Bridge 움직이는 책임 위임
